+++ plugins/module_utils/network/eos/facts/bgp_address_family/bgp_address_family.py
@@ -8,7 +8,7 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 """
-The eos bgp_af fact class
+The eos bgp_address_family fact class
 It is in this file the configuration is collected from the device
 for a given resource, parsed, and the facts tree is populated
 based on the configuration.
@@ -19,17 +19,17 @@ from copy import deepcopy
 from ansible_collections.ansible.netcommon.plugins.module_utils.network.common import (
     utils,
 )
-from ansible_collections.arista.eos.plugins.module_utils.network.eos.rm_templates.bgp_af import (
+from ansible_collections.arista.eos.plugins.module_utils.network.eos.rm_templates.bgp_address_family import (
     Bgp_afTemplate,
 )
-from ansible_collections.arista.eos.plugins.module_utils.network.eos.argspec.bgp_af.bgp_af import (
+from ansible_collections.arista.eos.plugins.module_utils.network.eos.argspec.bgp_address_family.bgp_address_family import (
     Bgp_afArgs,
 )
 import re
 
 
 class Bgp_afFacts(object):
-    """ The eos bgp_af facts class
+    """ The eos bgp_address_family facts class
     """
 
     def __init__(self, module, subspec="config", options="options"):
@@ -68,7 +68,7 @@ class Bgp_afFacts(object):
         if not data:
             data = self.get_config(connection)
 
-        # remove global configs from bgp_af
+        # remove global configs from bgp_address_family
         bgp_af_config = []
         vrf_set = ""
         start = False
@@ -102,13 +102,13 @@ class Bgp_afFacts(object):
                             af["network"], key=lambda k: k["address"]
                         )
 
-        ansible_facts["ansible_network_resources"].pop("bgp_af", None)
+        ansible_facts["ansible_network_resources"].pop("bgp_address_family", None)
 
         params = utils.remove_empties(
             utils.validate_config(self.argument_spec, {"config": objs})
         )
 
-        facts["bgp_af"] = params.get("config", [])
+        facts["bgp_address_family"] = params.get("config", [])
         ansible_facts["ansible_network_resources"].update(facts)
 
         return ansible_facts
