+++ plugins/module_utils/network/eos/facts/facts.py
@@ -65,7 +65,7 @@ from ansible_collections.arista.eos.plugins.module_utils.network.eos.facts.ospfv
 from ansible_collections.arista.eos.plugins.module_utils.network.eos.facts.ospf_interfaces.ospf_interfaces import (
     Ospf_interfacesFacts,
 )
-from ansible_collections.arista.eos.plugins.module_utils.network.eos.facts.bgp_af.bgp_af import (
+from ansible_collections.arista.eos.plugins.module_utils.network.eos.facts.bgp_address_family.bgp_address_family import (
     Bgp_afFacts,
 )
 
@@ -88,7 +88,7 @@ FACT_RESOURCE_SUBSETS = dict(
     ospfv2=Ospfv2Facts,
     ospfv3=Ospfv3Facts,
     ospf_interfaces=Ospf_interfacesFacts,
-    bgp_af=Bgp_afFacts,
+    bgp_address_family=Bgp_afFacts,
 )
