+++ plugins/module_utils/network/eos/rm_templates/bgp_address_family.py
