+++ tests/integration/targets/eos_bgp_address_family/defaults/main.yaml
