+++ tests/integration/targets/eos_bgp_address_family/tasks/cli.yaml
