+++ tests/integration/targets/eos_bgp_address_family/tests/common/_parsed.cfg
