+++ tests/integration/targets/eos_bgp_address_family/tests/common/gathered.yaml
@@ -1,6 +1,6 @@
 ---
 - debug:
-    msg: START eos_bgp_af gathered integration tests on connection={{ ansible_connection
+    msg: START eos_bgp_address_family gathered integration tests on connection={{ ansible_connection
       }}
 
 - include_tasks: _populate.yaml
@@ -10,7 +10,7 @@
     - name: Gathered the provided configuration with the exisiting running configuration
       become: true
       register: result
-      arista.eos.eos_bgp_af:
+      arista.eos.eos_bgp_address_family:
         config:
         state: gathered
