+++ tests/integration/targets/eos_bgp_address_family/tests/common/parsed.yaml
@@ -1,12 +1,12 @@
 ---
 - debug:
-    msg: START eos_bgp_af_interfaces parsed integration tests on connection={{ ansible_connection
+    msg: START eos_bgp_address_family_interfaces parsed integration tests on connection={{ ansible_connection
       }}
 
 - name: Provide the running configuration for parsing (config to be parsed)
   become: true
   register: result
-  arista.eos.eos_bgp_af:
+  arista.eos.eos_bgp_address_family:
     running_config: "{{ lookup('file', '_parsed.cfg') }}"
     state: parsed
