+++ tests/integration/targets/eos_bgp_address_family/vars/main.yaml
